+++ packages/terra-functional-testing/src/commands/utils/index.js
@@ -1,3 +1,4 @@
+const describeTests = require('./describeTests');
 const describeViewports = require('./describeViewports');
 const getViewports = require('./getViewports');
 const hideInputCaret = require('./hideInputCaret');
@@ -5,6 +6,7 @@ const setApplicationLocale = require('./setApplicationLocale');
 const setViewport = require('./setViewport');
 
 module.exports = {
+  describeTests,
   describeViewports,
   getViewports,
   hideInputCaret,